Make robustspecies wordsize-agnostic (#1555)
`robustspecies` built its result as `Int64[]` and stored it in
`NetworkProperties.robustspecies::Vector{Int}`. On 32-bit Julia
`Int == Int32`, so the assignment threw
`TypeError: in setfield!, expected Vector{Int32}, got Vector{Int64}`,
failing the CRN Theory tests in the "Modeling 32-bit" CI lane.



Agent-Harness: Claude Code 2.1.269
Agent-Model: claude-fable-5-1
Agent-Session: https://claude.ai/code/session_01KjwXuMxmYxyayQcn6tNwRo

+++ src/network_analysis.jl
@@ -1399,7 +1399,7 @@ function robustspecies(rn::ReactionSystem)
 
         # Find the complexes that do not belong to a terminal linkage class
         nonterminal_complexes = deleteat!([1:length(complexes);], vcat(tslcs...))
-        robust_species = Int64[]
+        robust_species = Int[]
 
         for (c_s, c_p) in Combinatorics.combinations(nonterminal_complexes, 2)
             # Check the difference of all the combinations of complexes. The support is the set of indices that are non-zero
